+++ config/autoload/unicaen-app.global.php
@@ -27,7 +27,7 @@ return [
             'nom'     => "EMC2",
             'desc'    => "Emploi Mobilité Carrière Compétences",
             'version' => "1.1.3",
-            'date'    => "08/04/2021",
+            'date'    => "12/04/2021",
             'contact' => ['mail' => "dsi.applications@unicaen.fr", /*'tel' => "01 02 03 04 05"*/],
             'mentionsLegales'        => "http://www.unicaen.fr/acces-direct/mentions-legales/",
             'informatiqueEtLibertes' => "http://www.unicaen.fr/acces-direct/informatique-et-libertes/",